refactor did pkg
