+++ did/model.go
@@ -4,11 +4,10 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/TBD54566975/ssi-sdk/crypto/jwx"
 	"github.com/multiformats/go-multibase"
-	"github.com/multiformats/go-multicodec"
 
 	"github.com/TBD54566975/ssi-sdk/crypto"
+	"github.com/TBD54566975/ssi-sdk/crypto/jwx"
 	"github.com/TBD54566975/ssi-sdk/cryptosuite"
 
 	"github.com/TBD54566975/ssi-sdk/util"
@@ -19,63 +18,7 @@ const (
 
 	// Base58BTCMultiBase Base58BTC https://github.com/multiformats/go-multibase/blob/master/multibase.go
 	Base58BTCMultiBase = multibase.Base58BTC
-
-	// Multicodec reference https://github.com/multiformats/multicodec/blob/master/table.csv
-
-	Ed25519MultiCodec   = multicodec.Ed25519Pub
-	X25519MultiCodec    = multicodec.X25519Pub
-	SECP256k1MultiCodec = multicodec.Secp256k1Pub
-	P256MultiCodec      = multicodec.P256Pub
-	P384MultiCodec      = multicodec.P384Pub
-	P521MultiCodec      = multicodec.P521Pub
-	RSAMultiCodec       = multicodec.RsaPub
-	SHA256MultiCodec    = multicodec.Sha2_256
 )
-
-// ResolutionResult encapsulates the tuple of a DID resolution https://www.w3.org/TR/did-core/#did-resolution
-type ResolutionResult struct {
-	Context            string `json:"@context,omitempty"`
-	ResolutionMetadata `json:"didResolutionMetadata,omitempty"`
-	Document           `json:"didDocument,omitempty"`
-	DocumentMetadata   `json:"didDocumentMetadata,omitempty"`
-}
-
-func (r *ResolutionResult) IsEmpty() bool {
-	if r == nil {
-		return true
-	}
-	return reflect.DeepEqual(r, ResolutionResult{})
-}
-
-// DocumentMetadata https://www.w3.org/TR/did-core/#did-document-metadata
-type DocumentMetadata struct {
-	Created       string `json:"created,omitempty" validate:"datetime"`
-	Updated       string `json:"updated,omitempty" validate:"datetime"`
-	Deactivated   bool   `json:"deactivated,omitempty"`
-	NextUpdate    string `json:"nextUpdate,omitempty"`
-	VersionID     string `json:"versionId,omitempty"`
-	NextVersionID string `json:"nextVersionId,omitempty"`
-	EquivalentID  string `json:"equivalentId,omitempty"`
-	CanonicalID   string `json:"canonicalId,omitempty"`
-}
-
-func (s *DocumentMetadata) IsValid() bool {
-	return util.NewValidator().Struct(s) == nil
-}
-
-// ResolutionError https://www.w3.org/TR/did-core/#did-resolution-metadata
-type ResolutionError struct {
-	Code                       string `json:"code"`
-	InvalidDID                 bool   `json:"invalidDid"`
-	NotFound                   bool   `json:"notFound"`
-	RepresentationNotSupported bool   `json:"representationNotSupported"`
-}
-
-// ResolutionMetadata https://www.w3.org/TR/did-core/#did-resolution-metadata
-type ResolutionMetadata struct {
-	ContentType string
-	Error       *ResolutionError
-}
 
 // Document is a representation of the did core specification https://www.w3.org/TR/did-core
 // TODO(gabe) enforce validation of DID syntax https://www.w3.org/TR/did-core/#did-syntax
