+++ did/pkh/testdata/did-pkh-bitcoin-doc.json
