+++ did/util_test.go
@@ -1,75 +1,14 @@
 package did
 
 import (
-	"context"
-	"crypto/ed25519"
 	"testing"
+
+	"github.com/mr-tron/base58"
+	"github.com/stretchr/testify/assert"
 
 	"github.com/TBD54566975/ssi-sdk/crypto"
 	"github.com/TBD54566975/ssi-sdk/crypto/jwx"
-	"github.com/mr-tron/base58"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
-
-func TestResolveKeyForDID(t *testing.T) {
-	resolver, err := NewResolver([]Resolver{KeyResolver{}}...)
-	require.NoError(t, err)
-	require.NotEmpty(t, resolver)
-
-	t.Run("empty resolver", func(tt *testing.T) {
-		_, err := ResolveKeyForDID(context.Background(), nil, "did:test", "test-kid")
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "resolver cannot be empty")
-	})
-
-	t.Run("empty did", func(tt *testing.T) {
-		_, err := ResolveKeyForDID(context.Background(), resolver, "", "test-kid")
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "not a valid did")
-	})
-
-	t.Run("unresolveable did", func(tt *testing.T) {
-		_, err := ResolveKeyForDID(context.Background(), resolver, "did:example:test", "test-kid")
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "unsupported method: example")
-	})
-
-	t.Run("unresolveable did", func(tt *testing.T) {
-		_, err := ResolveKeyForDID(context.Background(), resolver, "did:example:test", "test-kid")
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "unsupported method: example")
-	})
-
-	t.Run("invalid did", func(tt *testing.T) {
-		_, err := ResolveKeyForDID(context.Background(), resolver, "did:key:test", "test-kid")
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "could not expand did:key DID")
-	})
-
-	t.Run("valid did; no kid", func(tt *testing.T) {
-		_, didKey, err := GenerateDIDKey(crypto.Ed25519)
-		assert.NoError(tt, err)
-		_, err = ResolveKeyForDID(context.Background(), resolver, didKey.String(), "test-kid")
-		assert.Error(tt, err)
-		assert.Contains(tt, err.Error(), "has no verification methods with kid: test-kid")
-	})
-
-	t.Run("valid did; valid kid", func(tt *testing.T) {
-		privKey, didKey, err := GenerateDIDKey(crypto.Ed25519)
-		assert.NoError(tt, err)
-		expanded, err := didKey.Expand()
-		assert.NoError(tt, err)
-		kid := expanded.VerificationMethod[0].ID
-
-		key, err := ResolveKeyForDID(context.Background(), resolver, didKey.String(), kid)
-		assert.NoError(tt, err)
-		assert.NotEmpty(tt, key)
-
-		pubKey := privKey.(ed25519.PrivateKey).Public()
-		assert.Equal(tt, pubKey, key)
-	})
-}
 
 func TestGetKeyFromVerificationInformation(t *testing.T) {
 	t.Run("empty doc", func(tt *testing.T) {
@@ -241,18 +180,22 @@ func TestGetKeyFromVerificationInformation(t *testing.T) {
 		assert.Equal(t, pubKey, key)
 	})
 
-	t.Run("doc for did:key with multibase key", func(t *testing.T) {
-		privKey, didKey, err := GenerateDIDKey(crypto.Ed25519)
-		assert.NoError(t, err)
-		pubKey := privKey.(ed25519.PrivateKey).Public()
+	t.Run("doc for did with multibase key", func(t *testing.T) {
+		doc := Document{
+			ID: "did:key:z6MkiTBz1ymuepAQ4HEHYSF1H8quG5GLVVQR3djdX3mDooWp",
+			VerificationMethod: []VerificationMethod{
+				{
+					ID:              "#z6MkiTBz1ymuepAQ4HEHYSF1H8quG5GLVVQR3djdX3mDooWp",
+					Type:            "Ed25519VerificationKey2018",
+					Controller:      "did:key:z6MkiTBz1ymuepAQ4HEHYSF1H8quG5GLVVQR3djdX3mDooWp",
+					PublicKeyBase58: "4zvwRjXUKGfvwnParsHAS3HuSVzV5cA4McphgmoCtajS",
+				},
+			},
+		}
 
-		doc, err := didKey.Expand()
+		key, err := GetKeyFromVerificationMethod(doc, doc.VerificationMethod[0].ID)
 		assert.NoError(t, err)
-		assert.NotNil(t, doc)
-
-		key, err := GetKeyFromVerificationMethod(*doc, doc.VerificationMethod[0].ID)
-		assert.NoError(t, err)
-		assert.Equal(t, pubKey, key)
+		assert.NotEmpty(t, key)
 	})
 
 	t.Run("doc for did with JWK", func(t *testing.T) {
@@ -282,16 +225,4 @@ func TestGetKeyFromVerificationInformation(t *testing.T) {
 		assert.NoError(t, err)
 		assert.NotEmpty(t, key)
 	})
-}
-
-func TestEncodePublicKeyWithKeyMultiCodecType(t *testing.T) {
-	// unsupported type
-	_, err := encodePublicKeyWithKeyMultiCodecType(crypto.KeyType("unsupported"), nil)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "not a supported key type")
-
-	// bad public key
-	_, err = encodePublicKeyWithKeyMultiCodecType(crypto.Ed25519, nil)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "unknown public key type; could not convert to bytes")
 }
