+++ example/usecase/employer_university_flow/pkg/verifier.go
@@ -5,7 +5,8 @@ import (
 
 	"github.com/TBD54566975/ssi-sdk/credential"
 	"github.com/TBD54566975/ssi-sdk/crypto/jwx"
-	"github.com/TBD54566975/ssi-sdk/did"
+	"github.com/TBD54566975/ssi-sdk/did/resolver"
+
 	"github.com/pkg/errors"
 )
 
@@ -14,7 +15,7 @@ import (
 // 1. That the VP is valid
 // 2. All VCs in the VP are valid
 // 3. That the VC was issued by a trusted entity (implied by the presentation, according to the Presentation Definition)
-func ValidateAccess(verifier jwx.Verifier, resolver did.Resolver, submissionBytes []byte) error {
+func ValidateAccess(verifier jwx.Verifier, resolver resolver.Resolver, submissionBytes []byte) error {
 	_, _, vp, err := credential.VerifyVerifiablePresentationJWT(context.Background(), verifier, resolver, string(submissionBytes))
 	if err != nil {
 		return errors.Wrap(err, "failed to validate VP signature")
